requirements.txt
# ─── Dados e processamento ───────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
pyarrow==16.0.0

# ─── Machine Learning ────────────────────────────────────────────────────────
scikit-learn==1.5.0
imbalanced-learn==0.12.3
joblib==1.4.2

# ─── Experiment Tracking ─────────────────────────────────────────────────────
mlflow==2.13.2

# ─── API ─────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1

# ─── Monitoramento de Drift ──────────────────────────────────────────────────
evidently==0.4.30
streamlit==1.35.0

# ─── Visualização (EDA) ───────────────────────────────────────────────────────
matplotlib==3.9.0
seaborn==0.13.2

# ─── Testes ──────────────────────────────────────────────────────────────────
pytest==8.2.1
pytest-cov==5.0.0
httpx==0.27.0

# ─── Utilitários ─────────────────────────────────────────────────────────────
pyyaml==6.0.1
python-dotenv==1.0.1
loguru==0.7.2
